streamlit/requirements.txt
streamlit==1.5.0
requests==2.27.1
requests-toolbelt==0.9.1
Pillow==9.0.0
plotly==5.5.0